Impelementation of avps, base messages - only 2 more till we have all of them. many fixes and cleanups to API :]
git-svn-id: http://mobicents.googlecode.com/svn/trunk@1582 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/diameter/base/events/src/main/java/net/java/slee/resource/diameter/base/events/avp/DiameterAvp.java
@@ -45,7 +45,7 @@ public interface DiameterAvp extends Cloneable {
      * by IETF will have a Vendor-ID of zero.
      * @return the Vendor-ID, or zero if it is not specified.
      */
-    int getVendorID();
+    long getVendorID();
 
     /**
      * Return the AVP name, e.g., "Session-Id"
